# src/main/java/com/futurasmus/users_api/infrastructure/controller/UserController.java
package com.futurasmus.users_api.infrastructure.controller;

import java.net.URI;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.web.PageableDefault;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.ModelAttribute;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.futurasmus.users_api.application.dto.RequestUserDto;
import com.futurasmus.users_api.application.dto.RequestUserFilterDto;
import com.futurasmus.users_api.application.dto.RequestUserPatchDto;
import com.futurasmus.users_api.application.dto.ResponsePage;
import com.futurasmus.users_api.application.dto.ResponseUserDto;
import com.futurasmus.users_api.application.service.UserService;

import jakarta.validation.Valid;

@RestController
@RequestMapping("/api/users")
public class UserController {
    
    @Autowired
    private UserService userService;

    // CREATE
    @PostMapping
    public ResponseEntity<ResponseUserDto> createUser(@Valid @RequestBody RequestUserDto user) {
        ResponseUserDto saved = userService.createUser(user);
        return ResponseEntity.created(URI.create("/api/users/" + saved.id()))
            .contentType(MediaType.APPLICATION_JSON)
            .body(saved);
    }

    // READ
    @GetMapping
    public ResponseEntity<ResponsePage<ResponseUserDto>> getAllUsers(@ModelAttribute RequestUserFilterDto filter,
                                                     @PageableDefault(size = 10, sort = "id") Pageable pageable) {
        Page<ResponseUserDto> page = userService.getAllUsers(filter, pageable);
        return ResponseEntity.ok()
            .contentType(MediaType.APPLICATION_JSON)
            .body(new ResponsePage<>(page));
    }
    
    @GetMapping("/{userId}")
    public ResponseEntity<ResponseUserDto> getUserById(@PathVariable Long userId) {
        ResponseUserDto user = userService.getUserById(userId);
        return ResponseEntity.ok()
            .contentType(MediaType.APPLICATION_JSON)
            .body(user);
    }
    
    // UPDATE
    @PutMapping("/{userId}")
    public ResponseEntity<ResponseUserDto> updateUser(@PathVariable Long userId, @Valid @RequestBody RequestUserDto userDto) {
        ResponseUserDto updated = userService.updateUser(userId, userDto);
        return ResponseEntity.ok()
            .contentType(MediaType.APPLICATION_JSON)
            .body(updated);
    }

    @PatchMapping("/{userId}")
    public ResponseEntity<ResponseUserDto> updateUserPartial(@PathVariable Long userId, @Valid @RequestBody RequestUserPatchDto userDto) {
        ResponseUserDto updated = userService.updateUserPartial(userId, userDto);
        return ResponseEntity.ok()
            .contentType(MediaType.APPLICATION_JSON)
            .body(updated);
    }

    // DELETE
    @DeleteMapping("/{userId}")
    public ResponseEntity<Void> deleteUser(@PathVariable Long userId) {
        userService.deleteUser(userId);
        return ResponseEntity.noContent().build();
    }
    
}